Properly add test for large payloads
