--- packages/test/test-end-to-end-tests/src/test/payloadSize.spec.ts
@@ -1,95 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import { strict as assert } from "assert";
-import { SharedMap } from "@fluidframework/map";
-import { requestFluidObject } from "@fluidframework/runtime-utils";
-import {
-    ITestFluidObject,
-    ChannelFactoryRegistry,
-    ITestObjectProvider,
-    ITestContainerConfig,
-    DataObjectFactoryType,
-} from "@fluidframework/test-utils";
-import { describeNoCompat, itExpects } from "@fluidframework/test-version-utils";
-import { IContainer, IErrorBase } from "@fluidframework/container-definitions";
-import { ConfigTypes, IConfigProviderBase } from "@fluidframework/telemetry-utils";
-import { FlushMode } from "@fluidframework/runtime-definitions";
-
-describeNoCompat("Payload size", (getTestObjectProvider) => {
-    const mapId = "mapId";
-    const registry: ChannelFactoryRegistry = [
-        [mapId, SharedMap.getFactory()],
-    ];
-    const testContainerConfig: ITestContainerConfig = {
-        fluidDataObjectType: DataObjectFactoryType.Test,
-        registry,
-    };
-
-    let provider: ITestObjectProvider;
-    beforeEach(() => {
-        provider = getTestObjectProvider();
-    });
-    afterEach(async () => provider.reset());
-
-    let container1: IContainer;
-    let dataObject1: ITestFluidObject;
-    let dataObject2: ITestFluidObject;
-    let dataObject1map: SharedMap;
-    let dataObject2map: SharedMap;
-
-    const configProvider = ((settings: Record<string, ConfigTypes>): IConfigProviderBase => ({
-        getRawConfig: (name: string): ConfigTypes => settings[name],
-    }));
-
-    const setupContainers = async (
-        containerConfig: ITestContainerConfig,
-        featureGates: Record<string, ConfigTypes> = {},
-    ) => {
-        const configWithFeatureGates = {
-            ...containerConfig,
-            loaderProps: { configProvider: configProvider(featureGates) }
-        };
-
-        container1 = await provider.makeTestContainer(configWithFeatureGates);
-        dataObject1 = await requestFluidObject<ITestFluidObject>(container1, "default");
-        dataObject1.context.containerRuntime.setFlushMode(FlushMode.TurnBased);
-        dataObject1map = await dataObject1.getSharedObject<SharedMap>(mapId);
-
-        // Load the Container that was created by the first client.
-        const container2 = await provider.loadTestContainer(testContainerConfig);
-        dataObject2 = await requestFluidObject<ITestFluidObject>(container2, "default");
-        dataObject2map = await dataObject2.getSharedObject<SharedMap>(mapId);
-
-        await provider.ensureSynchronized();
-    };
-
-    const generateStringOfSize = (sizeInBytes: number): string => new Array(sizeInBytes + 1).join("0");
-    const containerError = async (container: IContainer) =>
-        new Promise<IErrorBase | undefined>((resolve) => container.once("closed", (error) => { resolve(error); }));
-
-    it("Send one big op, smaller than the max op size", async () => {
-        await setupContainers(testContainerConfig, {});
-        // Max op size is 768000, round down to account for some overhead
-        const largeString = generateStringOfSize(750000);
-        dataObject1map.set("key", largeString);
-
-        await provider.ensureSynchronized();
-
-        assert.strictEqual(dataObject2map.get("key"), largeString);
-    });
-
-    itExpects("Send one big op, larger than the max op size", [
-        { eventName: "fluid:telemetry:Container:ContainerClose" },
-    ], async () => {
-        await setupContainers(testContainerConfig, {});
-        const errorEvent = containerError(container1);
-
-        const largeString = generateStringOfSize(800000);
-        dataObject1map.set("key", largeString);
-
-        await errorEvent;
-    });
-});
